requirements.txt
discord
pyxivapi
python-dotenv
